requirements.txt
albumentations>=0.4.3
diffusers
opencv-python>=4.1.2.30
pudb>=2019.2
invisible-watermark
imageio>=2.9.0
imageio-ffmpeg>=0.4.2
pytorch-lightning>=1.4.2,<2.0.0
omegaconf>=2.1.1
hydra-core>=1.3.2
test-tube>=0.7.5
streamlit>=0.73.1
einops>=0.3.0
torch-fidelity>=0.3.0
transformers>=4.19.2
torchmetrics>=0.6.0
kornia>=0.6
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
-e git+https://github.com/openai/CLIP.git@main#egg=clip
git+https://github.com/NVlabs/nvdiffrast.git
-e git+https://github.com/CompVis/stable-diffusion.git#egg=latent-diffusion
ipykernel
jsonpickle
Pillow
lark
altair
ipympl
ipython
ipython-genutils
ipywidgets
jupyterlab
jupyterlab-widgets
ninja
sca2d
scikit-image
scikit-learn
scikit-video
scipy
sympy
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardx
trimesh
stackview
typeguard
fire
